Commit 5

+++ domain/src/main/java/com/coherentsolutions/domain/Product.java
@@ -4,7 +4,7 @@ public class Product {
     private double rate;
     private double price;
 
-    public Product(String name, int rate, int price) {
+    public Product(String name, double rate, double price) {
         this.name = name;
         this.rate = rate;
         this.price = price;
